refact: extract socket handlers from server/index.ts
