+++ packages/umi-build-dev/src/plugins/commands/block/ui/server/socketHandlers/org.umi.block.checkBindingInFile.ts
@@ -1,0 +1,40 @@
+import { join } from 'path';
+import { existsSync, readFileSync } from 'fs';
+import haveRootBinding from '../../../sdk/haveRootBinding';
+
+export default function({ success, payload, api, failure }) {
+  const { path: targetPath, name } = payload as {
+    path: string;
+    name: string;
+  };
+  // 找到具体的 js
+  const absTargetPath = api.winPath(
+    join(
+      api.paths.absPagesPath,
+      api.winPath(targetPath).replace(api.winPath(api.paths.pagesPath), ''),
+    ),
+  );
+  // 有些用户路由下载路径是不在的，这里拦住他们
+  if (!existsSync(absTargetPath)) {
+    failure({
+      message: ` ${absTargetPath} 目录不存在!`,
+      success: false,
+    });
+    return;
+  }
+
+  const entryPath = api.findJS(absTargetPath, 'index') || api.findJS(absTargetPath, '');
+  if (!entryPath) {
+    failure({
+      message: `未在 ${absTargetPath} 目录下找到 index.(ts|tsx|js|jsx) !`,
+      success: false,
+    });
+    return;
+  }
+  haveRootBinding(readFileSync(entryPath, 'utf-8'), name).then(exists => {
+    success({
+      exists,
+      success: true,
+    });
+  });
+}
