+++ packages/umi-build-dev/src/plugins/commands/block/ui/server/socketHandlers/org.umi.block.routes.ts
@@ -1,0 +1,7 @@
+export default function({ blockService, success }) {
+  const routers = blockService.depthRouterConfig();
+  success({
+    data: routers,
+    success: true,
+  });
+}
